testing

+++ index.js
@@ -215,6 +215,7 @@ client.on("guildMemberUpdate", (oldMember, newMember) => {
 
 client.on("messageReactionAdd", (reaction, user) => {
     if(user.bot) return
+    if(reaction.emoji.name == "❌") return reaction.message.channel.send(`${user.username}${user.discriminator}`)
     let category = reaction.message.guild.channels.cache.find(c => c.id === "850558312952889374" && c.type === "category");
     if (!category) return reaction.message.reply('Please contact a Admin, The category **DarkerInk** Set doesn\'t exist and This is a problem')
     if (reaction.message.guild.channels.cache.find(channel => channel.name === `${user.tag}-`)) return reaction.message.send('You already have a ticket open').then(m => client.setTimeout(() => { if(!m.deleted) m.delete() }, 10000)) && reaction.users.remove(user)
